added cleaned files
Listings is the cleaned data of listings after changing all the true and false to 1 and 0 respectively. 
Also deleted all the unnecessary columns.
In sentiment.csv I ran polarity and sentiment analysis from text blob to the reviews data of and created a new file with those values. Then I removed all the blank spaces and deleted the rows with no comments.